chore: move field_textinput_removable.js into src
