+++ blocks_vertical/procedures.js
@@ -24,7 +24,7 @@
 
 import * as Blockly from "blockly/core";
 import { Colours } from "../src/colours.js";
-import { FieldTextInputRemovable } from "../core/field_textinput_removable.js";
+import { FieldTextInputRemovable } from "../src/fields/field_textinput_removable.js";
 
 class DuplicateOnDragDraggable {
   constructor(block) {
